Remove class definition and rename functions to fit launch core naming standards
